console-app/requirements.txt
huggingface-hub>=0.14.1
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
